==> 2. Data Preprocessing.ipynb <==
# imports
import pandas as pd
import matplotlib.pyplot as plt
import statistics

# read oti csv file in
file_path_oti = 'D:\\sentences_sonar\\sentences_sonar_om.csv'
df_oti = pd.read_csv(file_path_oti)

# read ti csv file in
file_path_ti = 'D:\\sentences_sonar\\sentences_sonar_te.csv'
df_ti = pd.read_csv(file_path_ti)
df_ti['Category'] = 'ti'

# read oti csv file in
file_path_niet = 'D:\\sentences_sonar\\sentences_sonar_niet.csv'
df_niet = pd.read_csv(file_path_niet)

print('number of oti sentences: ', df_oti.shape)
print('number of ti sentences: ', df_ti.shape)
print('number of niet sentences: ', df_niet.shape)

# Filtering for Verbs

### Counting Verbs

# get the verbs from the different categories
oti_verbs = df_oti['Verb']
ti_verbs = df_ti['Verb']

# see which verbs are more than 10 times in oti and ti
oti_verb_count = {}
for v in oti_verbs:
    if v in oti_verb_count.keys():
        oti_verb_count[v] += 1
    else:
        oti_verb_count[v] = 1

ti_verb_count = {}
for v in ti_verbs:
    if v in ti_verb_count.keys():
        ti_verb_count[v] += 1
    else:
        ti_verb_count[v] = 1

verbs_tenplus_oti = []
for k in oti_verb_count:
    if oti_verb_count[k] > 10:
        verbs_tenplus_oti.append(k)
verbs_tenplus_ti = []
for k in ti_verb_count:
    if ti_verb_count[k] > 10:
        verbs_tenplus_ti.append(k)

# add these verbs to a list
verb_list = []
for verb in verbs_tenplus_oti:
    if verb in verbs_tenplus_ti:
        verb_list.append(verb)

len(verb_list)

### Checking a few sentences per sentence

# show 2 sentences for each verb in the oti group
# for i in verb_list:
#     print(i)
#     for j in df_oti.loc[df_oti['Verb'] == i]['Sentence'].values[:2]:
#         print(j)
#     print('-' * 25)

# show 2 sentences for each verb in the ti group
# for i in verb_list:
#     print(i)
#     for j in df_ti.loc[df_ti['Verb'] == i]['Sentence'].values[:2]:
#         print(j)
#     print('-' * 25)

# show 2 sentences for each verb in the niet + ti group
# for i in verb_list:
#     print(i)
#     for j in df_niet.loc[df_niet['Verb'] == i]['Sentence'].values[:2]:
#         print(j)
#     print('-' * 25)

### Filtering

# filter the verbs to only get verbs with optional 'om' construction
verbs_to_filter = ['hebben', 'zijn', 'krijgen', 'maken', 'staan', 'worden', 'gaan', 'zullen', 'moeten', 'vinden', 'doen', 'willen', 'blijven', 'kunnen',
'zien', 'halen', 'nemen', 'geven', 'voelen', 'laten', 'lopen', 'liggen', 'vallen', 'zetten', 'schijnen', 'achten', 'in_zien', 'getuigen',
'onderstrepen', 'erkennen', 'beklemtonen', 'claimen', 'zweren', 'lijken', 'beginnen', 'komen', 'blijken', 'hoeven', 'zitten', 'denken', 'werken',
'mogen', 'weten', 'schrijven', 'zeggen', 'stoppen', 'durven', 'houden', 'stellen', 'bestaan', 'steunen', 'op_geven', 'herinneren', 'roepen',
'voor_hebben', 'beschuldigen', 'geloven', 'waarschuwen', 'wachten', 'antwoorden', 'ijveren', 'dreigen', 'melden', 'behoren', 'haten', 'toe_geven',
'voor_houden', 'bewijzen', 'bedoelen', 'zorgen', 'gelijk_hebben', 'herhalen', 'hangen', 'aan_geven', 'duren', 'verklaren', 'noemen', 'behoeven',
'vertellen', 'horen', 'menen', 'heten', 'af_spreken', 'bekennen', 'ontdekken', 'op_houden', 'toe_voegen', 'toe_komen', 'bewegen', 'op_gaan', 'drukken',
'beschouwen', 'aanvaarden', 'toe_wensen', 'op_merken', 'trachten', 'over_laten', 'vermoeden', 'verwijten', 'vermelden', 'aan_tonen', 'beweren',
'schreeuwen', 'veinzen', 'pretenderen', 'garanderen', 'pogen', 'bevestigen', 'plegen', 'verbazen', 'door_gaan', 'ontkennen', 'verkondigen',
'bestrijden', 'benadrukken', 'nodigen', 'aan_gaan', 'suggereren', 'betreuren', 'onthullen', 'voort_gaan', 'verzekeren', 'gebaren', 'mee_delen']

verb_list_filtered = [i for i in verb_list if i not in verbs_to_filter]

len(verb_list_filtered)

df_oti_filtered = df_oti.loc[df_oti['Verb'].isin(verb_list_filtered)]
df_ti_filtered = df_ti.loc[df_ti['Verb'].isin(verb_list_filtered)]
df_niet_filtered = df_niet.loc[df_niet['Verb'].isin(verb_list_filtered)]

print('number of oti sentences: ', df_oti_filtered.shape)
print('number of ti sentences: ', df_ti_filtered.shape)
print('number of niet sentences: ', df_niet_filtered.shape)

# Removing Overlap Between Groups

# checking overlap between oti and ti
filenames_overlap_oti_ti = set(df_oti_filtered['Filename']) & set(df_ti_filtered['Filename'])
len(filenames_overlap_oti_ti)

# checking overlap between ti and niet + ti
filenames_overlap_ti_niet = set(df_ti_filtered['Filename']) & set(df_niet_filtered['Filename'])
len(filenames_overlap_ti_niet)

# check overlap between oti and niet + ti
filenames_overlap_oti_niet = set(df_oti_filtered['Filename']) & set(df_niet_filtered['Filename'])
len(filenames_overlap_oti_niet)

# delete filenames_overlap_oti_ti from both oti and ti categories by filtering on files that are not in the list
df_oti_filtered = df_oti_filtered[~df_oti_filtered["Filename"].isin(filenames_overlap_oti_ti)] # ~ inverts the mask
df_ti_filtered = df_ti_filtered[~df_ti_filtered["Filename"].isin(filenames_overlap_oti_ti)]

# delete filenames_overlap_ti_niet from 'ti' category
df_ti_filtered = df_ti_filtered[~df_ti_filtered["Filename"].isin(filenames_overlap_ti_niet)] # ~ inverts the mask

# delete filenames_overlap_oti_niet from both oti and niet categories
df_oti_filtered = df_oti_filtered[~df_oti_filtered["Filename"].isin(filenames_overlap_oti_niet)] # ~ inverts the mask
df_niet_filtered = df_niet_filtered[~df_niet_filtered["Filename"].isin(filenames_overlap_oti_niet)]

print('number of oti sentences: ', df_oti_filtered.shape)
print('number of ti sentences: ', df_ti_filtered.shape)
print('number of niet sentences: ', df_niet_filtered.shape)

# Removing Overlap Within Groups

df_oti_filtered = df_oti_filtered.drop_duplicates(subset='Sentence')
df_ti_filtered = df_ti_filtered.drop_duplicates(subset='Sentence')
df_niet_filtered = df_niet_filtered.drop_duplicates(subset='Sentence')

# Adding Sentence Length and Removing Outliers

df_niet_filtered = df_niet_filtered.drop(labels='Unnamed: 0', axis='columns')

df_filtered = pd.concat([df_oti_filtered, df_ti_filtered, df_niet_filtered])

# add Sentence Length column to dataframe with the length of the sentence
df_filtered = df_filtered.reset_index()
for i in range(df_filtered.shape[0]):
    df_filtered.loc[i, 'Sentence_length'] = len(df_filtered['Sentence'][i].split())

Q1 = df_filtered['Sentence_length'].quantile(0.25)
Q3 = df_filtered['Sentence_length'].quantile(0.75)
IQR = Q3 - Q1
lower = Q1 - 1.5 * IQR
upper = Q3 + 1.5 * IQR

upper

df_no_outliers = df_filtered[(df_filtered['Sentence_length'] > lower) & (df_filtered['Sentence_length'] <= upper)]

df_no_outliers['Category'].value_counts()

# use the median of the sentence length as the cutoff between long and short sentences
cutoff = statistics.median(df_no_outliers['Sentence_length'])
print(cutoff)

# adding a short/long label to dataframe based on cutoff
df_no_outliers['Length Label'] = ['short' if x < cutoff else 'long' for x in df_no_outliers['Sentence_length']]

# Make All the Groups the Same Size

# make groups the same size (the size of the smallest group)
df_same_size = df_no_outliers.groupby('Category').nth[:min(df_no_outliers['Category'].value_counts())]

df_same_size['Category'].value_counts()

df_same_size['Length Label'].value_counts()

df_same_size[df_same_size['Category'] == 'oti']['Length Label'].value_counts()

df_same_size[df_same_size['Category'] == 'ti']['Length Label'].value_counts()

df_same_size[df_same_size['Category'] == 'niet + ti']['Length Label'].value_counts()

# Saving Filtered Sentences to CSV

df_same_size.to_csv("D:\\preprocessed_sentences.csv")
