==> 3. BERT pipelines.ipynb <==
# imports
import torch
from transformers import AutoTokenizer, AutoModel
import pandas as pd

# read in preprocessed sentences
txt = pd.read_csv('D:\\preprocessed_sentences.csv')[['Category', 'Sentence', 'Length Label']]
txt = list(txt.itertuples(index=False, name=None))

# word embeddings function
def word_embeddings(model, tokenizer, wrangled_txt):
    tokenized_txt = [(i[0], tokenizer.tokenize(i[1]), i[2]) for i in wrangled_txt]
    ids_tokens = [(i[0], tokenizer.convert_tokens_to_ids(i[1]), i[2]) for i in tokenized_txt]
    segment_ids = [(i[0], [1] * len(i[1]), i[2]) for i in tokenized_txt]
    token_tensor = [(i[0], torch.tensor([i[1]]), i[2]) for i in ids_tokens]
    segment_tensor = [(i[0], torch.tensor([i[1]]), i[2]) for i in segment_ids]
    hidden_states_list = []
    for i in range(len(token_tensor)):
        with torch.no_grad():
            outputs = model(token_tensor[i][1], segment_tensor[i][1])
        hidden_states = outputs[-1]
        hidden_states_list.append((token_tensor[i][0], hidden_states, token_tensor[i][2]))
    token_embeddings_list = []
    for h_s in hidden_states_list:
        token_embeddings = torch.stack(h_s[1], dim=0)
        token_embeddings = torch.squeeze(token_embeddings, dim=1) # remove the batch dimension
        token_embeddings = token_embeddings.permute(1,0,2)
        token_embeddings_list.append((h_s[0], token_embeddings, h_s[2]))
    word_embeddings = []
    for t_e in token_embeddings_list:
        token_vectors_sum = []
        for token in t_e[1]:
            sum_vector = torch.sum(token[-4:], dim=0) # sum the last 4 hidden layers
            token_vectors_sum.append(sum_vector)
        word_embeddings.append((t_e[0], token_vectors_sum, t_e[2]))
    return word_embeddings

# sentence embeddings function
def word_embedding_to_sentence_embedding(word_embedding):
    sentence_embedding = []
    for i in range(len(word_embedding[1][0])):
        total = 0
        for j in word_embedding[1]:
            total = torch.add(total, j[i])
        average = total / len(word_embedding[1])
        sentence_embedding.append(average.item())
    return sentence_embedding

# RobBERT-2023

# tokenizer, model
robbert_tokenizer = AutoTokenizer.from_pretrained("DTAI-KULeuven/robbert-2023-dutch-large")
robbert_model = AutoModel.from_pretrained("DTAI-KULeuven/robbert-2023-dutch-large", output_hidden_states = True, return_dict = False)
robbert_model.eval()

# data preprocessing
wrangled_txt = [(i[0], '<s> ' + i[1], i[2]) for i in txt]
len(wrangled_txt)

wrangled_txt_1 = wrangled_txt[:1000]
word_embeddings_robbert_1 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_1)
sentence_embeddings_robbert_1 = []
for word_embedding in word_embeddings_robbert_1:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_1.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_1 = pd.DataFrame(sentence_embeddings_robbert_1, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_1.to_csv('D:\\robbert_embeddings_1.csv')

wrangled_txt_2 = wrangled_txt[1000:2000]
word_embeddings_robbert_2 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_2)
sentence_embeddings_robbert_2 = []
for word_embedding in word_embeddings_robbert_2:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_2.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_2 = pd.DataFrame(sentence_embeddings_robbert_2, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_2.to_csv('D:\\robbert_embeddings_2.csv')

wrangled_txt_3 = wrangled_txt[2000:3000]
word_embeddings_robbert_3 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_3)
sentence_embeddings_robbert_3 = []
for word_embedding in word_embeddings_robbert_3:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_3.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_3 = pd.DataFrame(sentence_embeddings_robbert_3, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_3.to_csv('D:\\robbert_embeddings_3.csv')

wrangled_txt_4 = wrangled_txt[3000:4000]
word_embeddings_robbert_4 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_4)
sentence_embeddings_robbert_4 = []
for word_embedding in word_embeddings_robbert_4:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_4.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_4 = pd.DataFrame(sentence_embeddings_robbert_4, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_4.to_csv('D:\\robbert_embeddings_4.csv')

wrangled_txt_5 = wrangled_txt[4000:5000]
word_embeddings_robbert_5 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_5)
sentence_embeddings_robbert_5 = []
for word_embedding in word_embeddings_robbert_5:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_5.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_5 = pd.DataFrame(sentence_embeddings_robbert_5, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_5.to_csv('D:\\robbert_embeddings_5.csv')

wrangled_txt_6 = wrangled_txt[5000:6000]
word_embeddings_robbert_6 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_6)
sentence_embeddings_robbert_6 = []
for word_embedding in word_embeddings_robbert_6:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_6.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_6 = pd.DataFrame(sentence_embeddings_robbert_6, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_6.to_csv('D:\\robbert_embeddings_6.csv')

wrangled_txt_7 = wrangled_txt[6000:7000]
word_embeddings_robbert_7 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_7)
sentence_embeddings_robbert_7 = []
for word_embedding in word_embeddings_robbert_7:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_7.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_7 = pd.DataFrame(sentence_embeddings_robbert_7, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_7.to_csv('D:\\robbert_embeddings_7.csv')

wrangled_txt_8 = wrangled_txt[7000:]
word_embeddings_robbert_8 = word_embeddings(robbert_model, robbert_tokenizer, wrangled_txt_8)
sentence_embeddings_robbert_8 = []
for word_embedding in word_embeddings_robbert_8:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_robbert_8.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_robbert_8 = pd.DataFrame(sentence_embeddings_robbert_8, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_robbert_8.to_csv('D:\\robbert_embeddings_8.csv')

# BERTje

# tokenizer, model
bertje_tokenizer = AutoTokenizer.from_pretrained("GroNLP/bert-base-dutch-cased")
bertje_model = AutoModel.from_pretrained("GroNLP/bert-base-dutch-cased", output_hidden_states = True, return_dict = False)
bertje_model.eval()

# preprocessing data
wrangled_txt = [(i[0], '[CLS] ' + i[1] + ' [SEP]', i[2]) for i in txt]

wrangled_txt_1 = wrangled_txt[:1000]
word_embeddings_bertje_1 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_1)
sentence_embeddings_bertje_1 = []
for word_embedding in word_embeddings_bertje_1:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_1.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_1 = pd.DataFrame(sentence_embeddings_bertje_1, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_1.to_csv('D:\\bertje_embeddings_1.csv')

wrangled_txt_2 = wrangled_txt[1000:2000]
word_embeddings_bertje_2 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_2)
sentence_embeddings_bertje_2 = []
for word_embedding in word_embeddings_bertje_2:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_2.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_2 = pd.DataFrame(sentence_embeddings_bertje_2, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_2.to_csv('D:\\bertje_embeddings_2.csv')

wrangled_txt_3 = wrangled_txt[2000:3000]
word_embeddings_bertje_3 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_3)
sentence_embeddings_bertje_3 = []
for word_embedding in word_embeddings_bertje_3:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_3.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_3 = pd.DataFrame(sentence_embeddings_bertje_3, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_3.to_csv('D:\\bertje_embeddings_3.csv')

wrangled_txt_4 = wrangled_txt[3000:4000]
word_embeddings_bertje_4 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_4)
sentence_embeddings_bertje_4 = []
for word_embedding in word_embeddings_bertje_4:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_4.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_4 = pd.DataFrame(sentence_embeddings_bertje_4, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_4.to_csv('D:\\bertje_embeddings_4.csv')

wrangled_txt_5 = wrangled_txt[4000:5000]
word_embeddings_bertje_5 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_5)
sentence_embeddings_bertje_5 = []
for word_embedding in word_embeddings_bertje_5:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_5.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_5 = pd.DataFrame(sentence_embeddings_bertje_5, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_5.to_csv('D:\\bertje_embeddings_5.csv')

wrangled_txt_6 = wrangled_txt[5000:6000]
word_embeddings_bertje_6 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_6)
sentence_embeddings_bertje_6 = []
for word_embedding in word_embeddings_bertje_6:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_6.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_6 = pd.DataFrame(sentence_embeddings_bertje_6, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_6.to_csv('D:\\bertje_embeddings_6.csv')

wrangled_txt_7 = wrangled_txt[6000:7000]
word_embeddings_bertje_7 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_7)
sentence_embeddings_bertje_7 = []
for word_embedding in word_embeddings_bertje_7:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_7.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_7 = pd.DataFrame(sentence_embeddings_bertje_7, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_7.to_csv('D:\\bertje_embeddings_7.csv')

wrangled_txt_8 = wrangled_txt[7000:]
word_embeddings_bertje_8 = word_embeddings(bertje_model, bertje_tokenizer, wrangled_txt_8)
sentence_embeddings_bertje_8 = []
for word_embedding in word_embeddings_bertje_8:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_bertje_8.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_bertje_8 = pd.DataFrame(sentence_embeddings_bertje_8, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_bertje_8.to_csv('D:\\bertje_embeddings_8.csv')

# EuroBERT

# tokenizer, model
eurobert_tokenizer = AutoTokenizer.from_pretrained("EuroBERT/EuroBERT-210m")
eurobert_model = AutoModel.from_pretrained("EuroBERT/EuroBERT-210m", trust_remote_code=True, output_hidden_states = True, return_dict = False)
eurobert_model.eval()

# preprocessing data
wrangled_txt = [(i[0], '[CLS] ' + i[1] + ' [SEP]', i[2]) for i in txt]

wrangled_txt_1 = wrangled_txt[:1000]
word_embeddings_eurobert_1 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_1)
sentence_embeddings_eurobert_1 = []
for word_embedding in word_embeddings_eurobert_1:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_1.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_1 = pd.DataFrame(sentence_embeddings_eurobert_1, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_1.to_csv('D:\\eurobert_embeddings_1.csv')

wrangled_txt_2 = wrangled_txt[1000:2000]
word_embeddings_eurobert_2 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_2)
sentence_embeddings_eurobert_2 = []
for word_embedding in word_embeddings_eurobert_2:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_2.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_2 = pd.DataFrame(sentence_embeddings_eurobert_2, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_2.to_csv('D:\\eurobert_embeddings_2.csv')

wrangled_txt_3 = wrangled_txt[2000:3000]
word_embeddings_eurobert_3 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_3)
sentence_embeddings_eurobert_3 = []
for word_embedding in word_embeddings_eurobert_3:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_3.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_3 = pd.DataFrame(sentence_embeddings_eurobert_3, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_3.to_csv('D:\\eurobert_embeddings_3.csv')

wrangled_txt_4 = wrangled_txt[3000:4000]
word_embeddings_eurobert_4 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_4)
sentence_embeddings_eurobert_4 = []
for word_embedding in word_embeddings_eurobert_4:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_4.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_4 = pd.DataFrame(sentence_embeddings_eurobert_4, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_4.to_csv('D:\\eurobert_embeddings_4.csv')

wrangled_txt_5 = wrangled_txt[4000:5000]
word_embeddings_eurobert_5 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_5)
sentence_embeddings_eurobert_5 = []
for word_embedding in word_embeddings_eurobert_5:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_5.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_5 = pd.DataFrame(sentence_embeddings_eurobert_5, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_5.to_csv('D:\\eurobert_embeddings_5.csv')

wrangled_txt_6 = wrangled_txt[5000:6000]
word_embeddings_eurobert_6 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_6)
sentence_embeddings_eurobert_6 = []
for word_embedding in word_embeddings_eurobert_6:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_6.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_6 = pd.DataFrame(sentence_embeddings_eurobert_6, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_6.to_csv('D:\\eurobert_embeddings_6.csv')

wrangled_txt_7 = wrangled_txt[6000:7000]
word_embeddings_eurobert_7 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_7)
sentence_embeddings_eurobert_7 = []
for word_embedding in word_embeddings_eurobert_7:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_7.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_7 = pd.DataFrame(sentence_embeddings_eurobert_7, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_7.to_csv('D:\\eurobert_embeddings_7.csv')

wrangled_txt_8 = wrangled_txt[7000:]
word_embeddings_eurobert_8 = word_embeddings(eurobert_model, eurobert_tokenizer, wrangled_txt_8)
sentence_embeddings_eurobert_8 = []
for word_embedding in word_embeddings_eurobert_8:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_eurobert_8.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_eurobert_8 = pd.DataFrame(sentence_embeddings_eurobert_8, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_eurobert_8.to_csv('D:\\eurobert_embeddings_8.csv')

# mBERT

# word embeddings mBERT
mbert_tokenizer = AutoTokenizer.from_pretrained('bert-base-multilingual-cased')
mbert_model = AutoModel.from_pretrained("bert-base-multilingual-cased", output_hidden_states = True, return_dict = False)
mbert_model.eval()

wrangled_txt = [(i[0], '[CLS] ' + i[1] + ' [SEP]', i[2]) for i in txt]

wrangled_txt_1 = wrangled_txt[:1000]
word_embeddings_mbert_1 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_1)
sentence_embeddings_mbert_1 = []
for word_embedding in word_embeddings_mbert_1:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_1.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_1 = pd.DataFrame(sentence_embeddings_mbert_1, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_1.to_csv('D:\\mbert_embeddings_1.csv')

wrangled_txt_2 = wrangled_txt[1000:2000]
word_embeddings_mbert_2 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_2)
sentence_embeddings_mbert_2 = []
for word_embedding in word_embeddings_mbert_2:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_2.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_2 = pd.DataFrame(sentence_embeddings_mbert_2, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_2.to_csv('D:\\mbert_embeddings_2.csv')

wrangled_txt_3 = wrangled_txt[2000:3000]
word_embeddings_mbert_3 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_3)
sentence_embeddings_mbert_3 = []
for word_embedding in word_embeddings_mbert_3:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_3.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_3 = pd.DataFrame(sentence_embeddings_mbert_3, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_3.to_csv('D:\\mbert_embeddings_3.csv')

wrangled_txt_4 = wrangled_txt[3000:4000]
word_embeddings_mbert_4 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_4)
sentence_embeddings_mbert_4 = []
for word_embedding in word_embeddings_mbert_4:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_4.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_4 = pd.DataFrame(sentence_embeddings_mbert_4, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_4.to_csv('D:\\mbert_embeddings_4.csv')

wrangled_txt_5 = wrangled_txt[4000:5000]
word_embeddings_mbert_5 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_5)
sentence_embeddings_mbert_5 = []
for word_embedding in word_embeddings_mbert_5:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_5.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_5 = pd.DataFrame(sentence_embeddings_mbert_5, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_5.to_csv('D:\\mbert_embeddings_5.csv')

wrangled_txt_6 = wrangled_txt[5000:6000]
word_embeddings_mbert_6 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_6)
sentence_embeddings_mbert_6 = []
for word_embedding in word_embeddings_mbert_6:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_6.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_6 = pd.DataFrame(sentence_embeddings_mbert_6, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_6.to_csv('D:\\mbert_embeddings_6.csv')

wrangled_txt_7 = wrangled_txt[6000:7000]
word_embeddings_mbert_7 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_7)
sentence_embeddings_mbert_7 = []
for word_embedding in word_embeddings_mbert_7:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_7.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_7 = pd.DataFrame(sentence_embeddings_mbert_7, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_7.to_csv('D:\\mbert_embeddings_7.csv')

wrangled_txt_8 = wrangled_txt[7000:]
word_embeddings_mbert_8 = word_embeddings(mbert_model, mbert_tokenizer, wrangled_txt_8)
sentence_embeddings_mbert_8 = []
for word_embedding in word_embeddings_mbert_8:
    sentence_embedding = word_embedding_to_sentence_embedding(word_embedding)
    sentence_embeddings_mbert_8.append((word_embedding[0], sentence_embedding, word_embedding[2]))
df_mbert_8 = pd.DataFrame(sentence_embeddings_mbert_8, columns=['Category', 'Sentence Embedding', 'Length Label'])
df_mbert_8.to_csv('D:\\mbert_embeddings_8.csv')
